# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TSLPRB API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/results/{result_id}": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Result",
        "operationId": "get_result_v1_results__result_id__get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/results/{result_id}/detail": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Detail",
        "operationId": "get_detail_v1_results__result_id__detail_get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/results/{result_id}/paper": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get Review Paper",
        "operationId": "get_review_paper_v1_results__result_id__paper_get",
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Result Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewPaperQuestionOut"
                  },
                  "title": "Response Get Review Paper V1 Results  Result Id  Paper Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LocalizedOptions": {
        "properties": {
          "en": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "En"
          },
          "te": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Te"
          }
        },
        "type": "object",
        "required": [
          "en",
          "te"
        ],
        "title": "LocalizedOptions"
      },
      "LocalizedText": {
        "properties": {
          "en": {
            "type": "string",
            "title": "En"
          },
          "te": {
            "type": "string",
            "title": "Te"
          }
        },
        "type": "object",
        "required": [
          "en",
          "te"
        ],
        "title": "LocalizedText"
      },
...
      "ResultActionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "sub": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "sub"
        ],
        "title": "ResultActionOut"
      },
      "ResultDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "testTitleN": {
            "type": "integer",
            "title": "Testtitlen"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "maxScore": {
            "type": "number",
            "title": "Maxscore"
          },
          "cutoffPct": {
            "type": "number",
            "title": "Cutoffpct"
          },
          "qualified": {
            "type": "boolean",
            "title": "Qualified"
          },
          "accuracyPct": {
            "type": "number",
            "title": "Accuracypct"
          },
          "avgSecondsPerQuestion": {
            "type": "number",
            "title": "Avgsecondsperquestion"
          },
          "negativeMarks": {
            "type": "number",
            "title": "Negativemarks"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "wrong": {
            "type": "integer",
            "title": "Wrong"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ResultActionOut"
            },
            "type": "array",
            "title": "Actions"
          },
          "review": {
            "items": {
              "$ref": "#/components/schemas/ReviewRowOut"
            },
            "type": "array",
            "title": "Review"
          }
        },
        "type": "object",
        "required": [
          "id",
          "testTitleN",
          "title",
          "score",
          "maxScore",
          "cutoffPct",
          "qualified",
          "accuracyPct",
          "avgSecondsPerQuestion",
          "negativeMarks",
          "correct",
          "wrong",
          "skipped",
          "actions",
          "review"
        ],
        "title": "ResultDetailOut"
      },
      "ResultOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "test_id": {
            "type": "string",
            "title": "Test Id"
          },
          "attempt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score"
          },
          "cutoff": {
            "type": "number",
            "title": "Cutoff"
          },
          "qualified": {
            "type": "boolean",
            "title": "Qualified"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "accuracy": {
            "type": "number",
            "title": "Accuracy"
          },
          "per_section": {
            "items": {
              "$ref": "#/components/schemas/SectionScore"
            },
            "type": "array",
            "title": "Per Section"
          },
          "wrong": {
            "items": {
              "$ref": "#/components/schemas/WrongAnswer"
            },
            "type": "array",
            "title": "Wrong"
          }
        },
        "type": "object",
        "required": [
          "id",
          "test_id",
          "score",
          "max_score",
          "cutoff",
          "qualified",
          "rank",
          "accuracy",
          "per_section",
          "wrong"
        ],
        "title": "ResultOut"
      },
      "ReviewPaperQuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "section": {
            "type": "string",
            "enum": [
              "arithmetic",
              "reasoning",
              "gs",
              "telangana",
              "english"
            ],
            "title": "Section"
          },
          "text": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "options": {
            "$ref": "#/components/schemas/LocalizedOptions"
          },
          "avgSeconds": {
            "type": "number",
            "title": "Avgseconds"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "explanation": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object",
        "required": [
          "id",
          "section",
          "text",
          "options",
          "avgSeconds",
          "correct",
          "explanation"
        ],
        "title": "ReviewPaperQuestionOut"
      },
      "ReviewRowOut": {
        "properties": {
          "questionNo": {
            "type": "integer",
            "title": "Questionno"
          },
          "your": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your"
          },
          "seconds": {
            "type": "number",
            "title": "Seconds"
          }
        },
        "type": "object",
        "required": [
          "questionNo",
          "your",
          "seconds"
        ],
        "title": "ReviewRowOut"
      },
...
      "SectionScore": {
        "properties": {
          "section_id": {
            "type": "string",
            "title": "Section Id"
          },
          "name": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "wrong": {
            "type": "integer",
            "title": "Wrong"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "marks": {
            "type": "number",
            "title": "Marks"
          }
        },
        "type": "object",
        "required": [
          "section_id",
          "name",
          "correct",
          "wrong",
          "skipped",
          "marks"
        ],
        "title": "SectionScore"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WrongAnswer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "text": {
            "$ref": "#/components/schemas/LocalizedText"
          },
          "options": {
            "$ref": "#/components/schemas/LocalizedOptions"
          },
          "your_choice": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your Choice"
          },
          "correct_choice": {
            "type": "integer",
            "title": "Correct Choice"
          },
          "explanation": {
            "$ref": "#/components/schemas/LocalizedText"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "text",
          "options",
          "your_choice",
          "correct_choice",
          "explanation"
        ],
        "title": "WrongAnswer"
      }
    }
  }
}